+++ src/index.js
@@ -1,7 +1,9 @@
 const { PORT } = require("./config");
 
+
 const express = require("express")
 const swaggerUi = require('swagger-ui-express')
+const path = require("path")
 const yaml = require('yamljs')
 
 const app = express()
@@ -12,19 +14,25 @@ app.use(express.json())
 const index = require("./routes/index")
 const login = require("./routes/login")
 const signup = require("./routes/signup")
-const verification = require("./routes/verification")
+const profile = require("./routes/profile")
+const verification = require("./routes/verification");
 const swaggerDocument = yaml.load('spec.yaml')
+
+app.get("/", (req, res) => {
+    return res.sendFile(path.join(__dirname, "index.html"))
+})
 
 app.use("/", index)
 app.use("/login", login)
 app.use("/signup", signup)
+app.use("/profile", profile)
 app.use("/verification", verification)
 app.use('/api/docs', swaggerUi.serve, swaggerUi.setup(swaggerDocument))
 
 //404 error
 app.use((req, res, next) => {
     let ip = req.headers["x-forwarded-for"] || req.socket.remoteAddress
-    console.log(ip)
+    console.log("404 Not found!")
     return res.status(404).send("URL not found")
 })